+++ src/main/java/spacewars/model/game/elements/bullets/BulletNormalInvader.java
@@ -2,7 +2,7 @@ package spacewars.model.game.elements.bullets;
 
 public class BulletNormalInvader extends Bullet
 {
-    public BulletNormalInvader(int x, int y)
+    public BulletNormalInvader(double x, double y)
     {
         super(x, y, 10);
     }
